Changed some values to float, for more precision

+++ include/SSVBloodshed/Components/OBCHealth.h
@@ -13,28 +13,28 @@ namespace ob
 	class OBCHealth : public sses::Component
 	{
 		private:
-			int health, maxHealth;
+			float health, maxHealth;
 			ssvs::Ticker cooldown{1.f};
 
 		public:
 			ssvu::Delegate<void()> onDamage, onHeal;
 
-			OBCHealth(int mHealth) : health{mHealth}, maxHealth{mHealth} { }
+			OBCHealth(float mHealth) : health{mHealth}, maxHealth{mHealth} { }
 
 			inline void update(float mFT) override { if(cooldown.update(mFT)) cooldown.stop(); }
 
-			inline void heal(int mHealth) noexcept { health += mHealth; ssvu::clampMax(health, maxHealth); onHeal(); }
-			inline bool damage(int mDamage) noexcept
+			inline void heal(float mHealth) noexcept { health += mHealth; ssvu::clampMax(health, maxHealth); onHeal(); }
+			inline bool damage(float mDamage) noexcept
 			{
 				if(cooldown.isRunning()) return false;
 
 				cooldown.restart();
-				health = ssvu::getClampedMin(health - mDamage, 0);
+				health = ssvu::getClampedMin(health - mDamage, 0.f);
 				onDamage();
 				return true;
 			}
-			inline void setHealth(int mHealth) noexcept		{ health = ssvu::getClamped(mHealth, 0, maxHealth); }
-			inline void setMaxHealth(int mValue) noexcept	{ maxHealth = mValue; }
+			inline void setHealth(float mHealth) noexcept	{ health = ssvu::getClamped(mHealth, 0.f, maxHealth); }
+			inline void setMaxHealth(float mValue) noexcept	{ maxHealth = mValue; }
 			inline void setCooldown(float mValue) noexcept	{ cooldown.restart(mValue); }
 
 			inline bool isDead() const noexcept			{ return health <= 0; }
